# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rotaract South Asia Analytics API",
    "version": "1.1.0",
    "description": "Public, high-performance REST API and Model Context Protocol (MCP) server providing open analytics data for Rotaract and Rotary across South Asia (Zones 4, 5, 6, and 7; 44 districts; 2,870+ clubs).\n\n### Model Context Protocol (MCP) Integration\nThis service includes a Model Context Protocol (MCP) server for direct use by AI assistants (Cursor, Claude Desktop, Antigravity, VS Code):\n- **Local Stdio Command:** `npm run mcp` or `node scripts/mcp_server.js`\n- **Remote HTTP Endpoint:** `POST https://insights.rsmda.org/api/mcp` (JSON-RPC 2.0)\n- **Available AI Tools (14 tools):** `get_summary`, `get_leaderboards`, `get_districts`, `get_district_insights`, `get_zone_summary`, `search_clubs`, `get_club_profile`, `get_new_clubs`, `find_compliance_risks`, `find_dual_risk_clubs`, `get_interact_analytics`, `find_rotary_opportunities`, `get_foundation_giving`, `get_worldwide_rankings`.\n\nAll endpoints are open, unauthenticated, and CORS-enabled.",
    "contact": {
      "name": "Rotaract South Asia Analytics Team",
      "url": "https://insights.rsmda.org"
    },
    "license": {
      "name": "Open Data Commons Open Database License (ODbL)",
      "url": "https://opendatacommons.org/licenses/odbl/"
    }
  },
  "servers": [
    {
      "url": "https://insights.rsmda.org",
      "description": "Production server (Primary)"
    },
    {
...
    },
    {
...
    }
...
  "paths": {
...
    "/api/v1/worldwide": {
      "get": {
        "tags": [
          "Worldwide"
        ],
        "summary": "Worldwide Rotaract & Interact Statistics",
        "description": "Global totals, country growth leaderboards, district rankings, and new club charter statistics. Supports sorting and filtering to find top-performing districts globally.",
        "operationId": "getWorldwideStats",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Focus dataset: 'summary', 'country', 'district', 'interact', 'new_clubs', or 'all'",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "summary",
                "country",
                "district",
                "interact",
                "new_clubs"
              ],
              "default": "all"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Filter by country name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Filter by Rotary Zone (e.g. '6', 'Zone 6', '4,5,6,7', or 'South Asia')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "description": "Filter by geographic region ('south_asia', 'southeast_asia', 'pacific', 'europe', 'north_america', 'latin_america', or 'worldwide')",
            "schema": {
              "type": "string",
              "enum": [
                "south_asia",
                "southeast_asia",
                "pacific",
                "europe",
                "north_america",
                "latin_america",
                "worldwide"
              ],
              "default": "worldwide"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort metric: 'member_growth_pct', 'member_growth_abs', 'members', 'clubs', 'club_growth_pct', or 'district'",
            "schema": {
              "type": "string",
              "enum": [
                "member_growth_pct",
                "member_growth_abs",
                "members",
                "clubs",
                "club_growth_pct",
                "club_growth_abs",
                "district"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order: 'desc' or 'asc'",
            "schema": {
              "type": "string",
              "enum": [
                "desc",
                "asc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "minMembers",
            "in": "query",
            "required": false,
            "description": "Minimum reported members filter",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Limit number of records returned",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of records to skip for pagination",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worldwide statistics payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldwideStats"
                },
                "example": {
                  "total": 599,
                  "returned": 5,
                  "count": 5,
                  "offset": 0,
                  "limit": 5,
                  "scope": {
                    "region": "Worldwide",
                    "country": "All Countries",
                    "zone": "All Zones"
                  },
                  "sortBy": "member_growth_pct",
                  "sortOrder": "desc",
                  "districts": [
                    {
                      "rank": 1,
                      "worldwideRank": 1,
                      "southAsiaRank": 1,
                      "district": "3261",
                      "zone": 6,
                      "region": "South Asia",
                      "isSouthAsia": true,
                      "country": "India",
                      "activeClubs": 24,
                      "suspendedClubs": 0,
                      "totalMembers": 217,
                      "clubsGrowthAbs": 7,
                      "membersGrowthAbs": 119,
                      "clubsGrowthPct": 41.17647058823529,
                      "membersGrowthPct": 121.42857142857142
                    },
                    {
                      "rank": 2,
                      "worldwideRank": 2,
                      "southAsiaRank": null,
                      "district": "1030",
                      "zone": 19,
                      "region": "Europe, Middle East & Africa",
                      "isSouthAsia": false,
                      "activeClubs": 3,
                      "suspendedClubs": 0,
                      "totalMembers": 44,
                      "clubsGrowthAbs": 1,
                      "membersGrowthAbs": 18,
                      "clubsGrowthPct": 50,
                      "membersGrowthPct": 69.23076923076923
                    },
                    {
                      "rank": 3,
                      "worldwideRank": 3,
                      "southAsiaRank": 2,
                      "district": "3141",
                      "zone": 4,
                      "region": "South Asia",
                      "isSouthAsia": true,
                      "country": "India",
                      "activeClubs": 128,
                      "suspendedClubs": 0,
                      "totalMembers": 3153,
                      "clubsGrowthAbs": 4,
                      "membersGrowthAbs": 1277,
                      "clubsGrowthPct": 3.225806451612903,
                      "membersGrowthPct": 68.07036247334754
                    },
                    {
                      "rank": 4,
                      "worldwideRank": 4,
                      "southAsiaRank": null,
                      "district": "1320",
                      "zone": 19,
                      "region": "Europe, Middle East & Africa",
                      "isSouthAsia": false,
                      "activeClubs": 1,
                      "suspendedClubs": 0,
                      "totalMembers": 21,
                      "clubsGrowthAbs": 0,
                      "membersGrowthAbs": 8,
                      "clubsGrowthPct": 0,
                      "membersGrowthPct": 61.53846153846154
                    },
                    {
                      "rank": 5,
                      "worldwideRank": 5,
                      "southAsiaRank": null,
                      "district": "9705",
                      "zone": 8,
                      "region": "Australia / New Zealand / Pacific",
                      "isSouthAsia": false,
                      "activeClubs": 2,
                      "suspendedClubs": 0,
                      "totalMembers": 34,
                      "clubsGrowthAbs": 0,
                      "membersGrowthAbs": 12,
                      "clubsGrowthPct": 0,
                      "membersGrowthPct": 54.54545454545454
                    }
                  ]
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WorldwideStats": {
        "type": "object",
        "properties": {
          "totalClubs": {
            "type": "integer",
            "example": 11453
          },
          "totalMembers": {
            "type": "integer",
            "example": 165430
          },
          "totalInteractClubs": {
            "type": "integer",
            "example": 14920
          },
          "countryData": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "districtData": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "zoneData": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}